Add the ability to use aws secrets with calls to kraken-lib
kraken-lib (formerly k2) is designed to pass AWS environment variables
for authentication and region selection. Add the ability to pass that
information through the krak8s pod.

To use this ability, add base64 encoded versions of the aws access key id
and secret to the values file passed to helm as documented in values.yaml.